+++ shared/src/com/danielvandenbrink/corena/GameObject.java
@@ -2,10 +2,10 @@ package com.danielvandenbrink.corena;
 
 public class GameObject {
     private final long uuid;
-    private int x;
-    private int y;
+    private float x;
+    private float y;
 
-    public GameObject(final long uuid, final int x, final int y) {
+    public GameObject(final long uuid, final float x, final float y) {
         this.uuid = uuid;
         this.x = x;
         this.y = y;
@@ -15,11 +15,11 @@ public class GameObject {
         return uuid;
     }
 
-    public int x() {
+    public float x() {
         return x;
     }
 
-    public int y() {
+    public float y() {
         return y;
     }
 }
